move mixnet contract interaction traits to shared location
