+++ Cargo.toml
@@ -206,7 +206,7 @@ edition = "2024"
 license = "Apache-2.0"
 rust-version = "1.87.0"
 readme = "README.md"
-version = "1.20.4"
+version = "1.20.5"
 
 [workspace.dependencies]
 addr = "0.15.6"
@@ -616,3 +616,4 @@ exit = "deny"
 panic = "deny"
 unimplemented = "deny"
 unreachable = "deny"
+
