+++ contracts/mixnet/Cargo.toml
@@ -41,6 +41,8 @@ bs58 = { workspace = true }
 serde = { workspace = true, default-features = false, features = ["derive"] }
 semver = { workspace = true }
 
+nym-crypto = { workspace = true, optional = true }
+
 
 [dev-dependencies]
 anyhow.workspace = true
@@ -56,7 +58,7 @@ nym-contracts-common-testing = { workspace = true }
 [features]
 default = []
 contract-testing = ["mixnet-contract-common/contract-testing"]
-testable-mixnet-contract = ["nym-contracts-common-testing"]
+testable-mixnet-contract = ["nym-contracts-common-testing", "nym-crypto", "nym-crypto/asymmetric", "nym-crypto/rand"]
 schema-gen = ["mixnet-contract-common/schema", "cosmwasm-schema"]
 
 [lints]
