+++ contracts/node-families/Cargo.toml
@@ -35,6 +35,7 @@ nym-mixnet-contract-common = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
+nym-crypto = { workspace = true, features = ["asymmetric"] }
 nym-mixnet-contract = { workspace = true, features = ["testable-mixnet-contract"] }
 nym-contracts-common-testing = { workspace = true }
